Introduce coordinates limit flag that controls the segmentation
The segmentation may or may not happen if the string is huge.
Segmentation is done when the EditPathCoordinates reached its limit.
This leaves the user to whether have accurate SES and edit distance by
setting the limit to constants::no_limit, or have less accurate SES and
edit distance by setting a limit. Edit distance calculation itself is
independent of this if the only thing calculated is edit distance, so
the edit_distance() function is always accurate.